+++ pandevice/device.py
@@ -118,6 +118,7 @@ class Vsys(VersionedPanObject):
         "objects.SecurityProfileGroup",
         "objects.CustomUrlCategory",
         "objects.LogForwardingProfile",
+        "objects.DynamicUserGroup",
         "policies.Rulebase",
         "network.EthernetInterface",
         "network.AggregateInterface",